cloud_functions/requirements.txt
functions_framework
google-auth
requests